AITA-39 Modified details and remove common syntax
